server/requirements.txt
python-dotenv>=0.18.0
fastapi>=0.85.0
uvicorn[standard]>=0.18.3
jinja2>=3.1.2
aiohttp>=3.8.1
aiomysql>=0.1.1
orjson>=3.8.1
lesscpy>=0.15.1
fastapi_utils>=0.2.1
email_validator>=1.3.0
toml>=0.10.2
rich>=12.6.0
aiofiles>=22.1.0
trio>=0.22.0
hypercorn[trio]>=0.14.3